+++ src/bide/core.cljs
@@ -133,13 +133,21 @@
 ;; --- Browser History Binding API
 
 (defn start!
-  "Starts the bide routing handling using the goog.History
-  api as browser history event watching mechanism.
+  "Starts the bide routing handling using the `goog.history.Html5History` API as
+  browser history event watching mechanism.
 
-  If you want use the html5 history instance providing
-  a factory function that returns the Html5History object
-  instance."
-  [router {:keys [on-navigate default html5?]
+  Accepts router and configuration map. Required configuration keys are
+  `:on-navigate` and `:default`. `:on-navigate` is a function that would be
+  called each time route is changed providing route id, params and query as
+  arguments. `:default` used as default route id when URL doesn't match any
+  route registered in router. Optional configuration keys are `:html5?` (`false`
+  by default) and `:html5history` (new `goog.history.Html5History` instance by
+  default). Passing anything that evaluates to logical false as value of
+  `:html5?` would configure history to use fragment to store token. Pass factory
+  function that returns instance of `goog.history.Html5History` to
+  `:html5history` when you need to do some customizations to history instance
+  used to manage history events."
+  [router {:keys [on-navigate default html5? html5history]
            :or {html5? false}
            :as opts}]
   (let [default (if (vector? default) default [default nil])]
@@ -154,12 +162,13 @@
                 (if (str/blank? token)
                   (or (apply resolve router default) "/")
                   token)))]
-      (let [history (if html5?
-                      (doto (Html5History.)
+      (let [html5history (if (fn? html5history) (html5history) (Html5History.))
+            history (if html5?
+                      (doto html5history
                         (.setPathPrefix "")
                         (.setUseFragment false)
                         (.setEnabled true))
-                      (doto (Html5History.)
+                      (doto html5history
                         (.setUseFragment true)
                         (.setEnabled true)))
             initial-token (-initial-token history)