+++ CHANGES.md
@@ -1,5 +1,13 @@
 # Changelog #
 
+## Version 1.5.0
+
+Date: 2017-04-25
+
+- Add the ability to provide customized instance of `goog.history.Html5History`
+  that would be used to manage history events.
+
+
 ## Version 1.4.0
 
 Date: 2017-01-12